+++ CHANGELOG.md
@@ -18,8 +18,7 @@ Please follow the instructions before upgrading Cumulus
 ### Added
 
 - **CUMULUS-4146**
-  - Added new non-null column `active` to Postgres `granules`
-    table.
+  - Added new  column `active` to Postgres `granules` table.
   - Added `active` property to `granule` record schema.
   - Created migration for `active` column and index for granules table